Module cortex_m_rtfm::examples

Examples

Modules

_0_zero_tasks
Minimal example with zero tasks

_1_one_task
An application with one task

_2_two_tasks
Two tasks running at the same priority with access to the same resource

_3_preemption
Two tasks running at different priorities with access to the same resource

_4_nested
Nesting claims and how the preemption threshold works

_5_late_resources
Demonstrates initialization of resources in init.

_6_safe_static_mut_ref
Safe creation of &'static mut references

_7_generics
Working with resources in a generic fashion

_8_full_syntax
A showcase of the app! macro syntax