Struct cortex_m_rtfm::Resource

pub struct Resource<T, C> { /* fields omitted */ }

A resource with ceiling C

Methods

impl<T, RC> Resource<T, C<RC>> where
RC: GreaterThanOrEqual<U0>,
RC: LessThanOrEqual<UMAX>,

const fn new(data: T) -> Self

Creates a new resource

impl<T, RC> Resource<T, C<RC>>

fn access<'cs, TP, SC>(
&'static self,
_priority: &P<TP>,
_current_ceiling: &'cs C<SC>
) -> Ref<'cs, T> where
RC: GreaterThanOrEqual<TP>,
SC: GreaterThanOrEqual<RC>,

Grants data race free and deadlock free access to the resource data

This operation is zero cost and doesn't impose any additional blocking.

Requirements

To access the resource data these conditions must be met:

The resource ceiling must be greater than or equal to the task priority

The system ceiling must be greater than or equal to the resource ceiling