Trait cortex_m_rtfm::Resource

pub unsafe trait Resource {
    type Ceiling;
    type Data: 'static + Send;
    fn borrow<'cs, P>(&'cs self, _p: &'cs Priority<P>) -> &'cs Self::Data
    where
        P: IsGreaterOrEqual<Self::Ceiling, Output = True>,
    { ... }

    fn borrow_mut<'cs, P>(
        &'cs mut self, 
        _p: &'cs Priority<P>
    ) -> &'cs mut Self::Data
    where
        P: IsGreaterOrEqual<Self::Ceiling, Output = True>,
    { ... }

    fn claim<'cs, R, F, P>(&self, _p: &mut Priority<P>, f: F) -> R
    where
        F: FnOnce(&Self::Data, &mut Priority<Maximum<P, Self::Ceiling>>) -> R,
        P: Max<Self::Ceiling> + Unsigned,
        Self::Ceiling: Unsigned,
    { ... }

    fn claim_mut<'cs, R, F, P>(&mut self, _p: &mut Priority<P>, f: F) -> R
    where
        F: FnOnce(&mut Self::Data, &mut Priority<Maximum<P, Self::Ceiling>>) -> R,
        P: Max<Self::Ceiling> + Unsigned,
        Self::Ceiling: Unsigned,
    { ... }
}

A resource shared between two or more tasks

Associated Types

type Ceiling

The priority ceiling of the resource

type Data: 'static + Send

The data protected by the resource

Provided Methods

fn borrow<'cs, P>(&'cs self, _p: &'cs Priority<P>) -> &'cs Self::Data where
P: IsGreaterOrEqual<Self::Ceiling, Output = True>,

Borrows the resource data for the span of the current priority

fn borrow_mut<'cs, P>(
&'cs mut self,
_p: &'cs Priority<P>
) -> &'cs mut Self::Data where
P: IsGreaterOrEqual<Self::Ceiling, Output = True>,

Mutably borrows the resource data for the span of the current priority

fn claim<'cs, R, F, P>(&self, _p: &mut Priority<P>, f: F) -> R where
F: FnOnce(&Self::Data, &mut Priority<Maximum<P, Self::Ceiling>>) -> R,
P: Max<Self::Ceiling> + Unsigned,
Self::Ceiling: Unsigned,

Creates a critical section, by raising the task priority, to access the resource data

fn claim_mut<'cs, R, F, P>(&mut self, _p: &mut Priority<P>, f: F) -> R where
F: FnOnce(&mut Self::Data, &mut Priority<Maximum<P, Self::Ceiling>>) -> R,
P: Max<Self::Ceiling> + Unsigned,
Self::Ceiling: Unsigned,

Creates a critical section, by raising the task priority, to mutably access the resource data