Crate cortex_m_rtfm

Real Time for The Masses: high performance, predictable, bare metal task scheduler

Features

Priority based scheduler implemented in hardware with minimal bookkeeping.

Doesn't need a memory allocator to operate

Re-exports

pub use cortex_m_rtfm_macros::app;

Modules

event_task
An event task, a task that starts in response to an event (interrupt source)

idle
The idle function

init
The initialization function

schedulable_task
An schedulable task

Structs

Priority
Token that represents the current priority level of a task

Traits

Resource
A resource shared between two or more tasks

ScheduleAfter
schedule_after interface

ScheduleNow
schedule_now interface

Functions

atomic
Executes the given closure atomically